requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
yt-dlp==2022.4.8
aiohttp==3.8.1
tenacity==8.0.1
psutil==5.9.0
PyYAML
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
SQLAlchemy==1.4.29
requests==2.26.0
